Restructures dropdown rendering
