+++ frontend/components/forms/fields/SelectTargetsInput/SelectTargetsMenu/helpers.js
