+++ frontend/components/forms/fields/SelectTargetsInput/TargetDetails/index.js
