+++ frontend/components/forms/fields/SelectTargetsInput/TargetOption/_styles.scss
