+++ frontend/components/forms/fields/SelectTargetsInput/index.js
@@ -1,0 +1,11 @@
+import Details from './TargetDetails';
+import Input from './SelectTargetsInput';
+import Option from './TargetOption';
+import Menu from './SelectTargetsMenu';
+
+export default {
+  Details,
+  Input,
+  Menu,
+  Option,
+};
