--- frontend/components/queries/SelectTargetsInput/index.js
@@ -1,1 +1,0 @@
-export default from './SelectTargetsInput';
